backend/requirements.txt
# ssec-Sentinel Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
aiohttp==3.9.1
pandas==2.1.3
numpy==1.26.2
geopandas==0.14.1
feedparser==6.0.10
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic==2.5.0
cachetools==5.3.2
loguru==0.7.2
pytest==7.4.3
black==23.12.0
requests==2.31.0
redis==5.0.1
python-multipart==0.0.6
